Refactor processing code
The generation code has been removed for the time being; focus now
is on getting the basic skeleton working.  Test cases are included
for many of the corner cases that I've identified so it's a bit easier
to remember what is and is not tolerated.  The initial/deriv/vars split
has been removed and processing is done on the full expression rather
than a set of disembodied lhs and rhs.