Support for extracting all parameters from the model
Works via a generated `contents` function, which extracts all parameters
from the struct.  It's pretty thick so will just get *everything* including
all the generated `dim_<x>` parameters and the like.  But that should be
helpful for debugging and it's not necessary really for this to be written
for speed.  For a model with 48 bits of stored data over 24 entries in the
struct this is ~2us so not too bad.

Note that if this is called before the model is initialised then junk will
be returned for all the uninitialised numbers.  I don't think I can deal with
that sensibly without storing a flag in the struct to indicate if it's ever
been initialised, but that might not be a bad idea.