Start parsing delay variables
Probably incomplete, but this pulls most of the information from delay
variables for use in the generation.  Includes a few more checks for
properties of delay variables.  Array delay models are going to be a bit
more complicated (involving more allocations and book-keeping) so I've
excluded them for now to get the simple case working.