Lots of tweaks to please R CMD check
We still get lots of warnings about registration; things like:

```
Registration problems:
  Evaluating ‘paste0(name, "_info")’ during check gives error
‘object 'name' not found’:
   .Call(paste0(name, "_info"), PACKAGE = dll)
  Evaluating ‘self$C$create’ during check gives error
‘object 'self' not found’:
   .Call(self$C$create, pars)
  Evaluating ‘self$C$init’ during check gives error
```

This is all from code that is generated, so really should not be
checked here.  There might be a reasonable hack way out of this, but
given it's only a `NOTE` and not even a `WARNING` I'm going to ignore
this but it's an issue for eventual CRAN submission I expect.