Support for constant derivatives
This is a corner case, but seems worth dealing with for completeness.
If

`deriv(y) <- <constant>`

then we need to be careful not to leave `deriv_y` marked as constant because
it leaves the derivative calculations incorrect (we get undefined values
in the dydts!).  Includes a simple test of a constant model, which really
nobody will ever run in real life.