Towards a nicer interface for pulling variables out
With array variables, the deSolve matrix-only approach is going to
break down quite badly.  This is especially acute for matrices and
3d arrays where we can't even escape this with a sensible numbering
system.

The idea here is to get the interface right and then if this is
a speed sink I'll write something (possibly more generated code) to get
it done faster.