Generator is now function with user pars as args
Rather than returning the R6 generator object, instead we return a function
that has been generated so that user parameters are the function parameters.
Required parameters are listed first, followed by ones with defaults (though
the defaults are not included - instead we store them as `NULL` but that
might change).

This is not perfect because the `set_user` branch doesn't get fixed -- that's
going to require some gymnastics to get working with R6.  I've fudged it
for now at least.  It also means that we don't have access to the generator
itself which might be a problem when generating packages.