Support for full user arrays
This hopefully completes the core support for user variables.  Now
it is valid to use:

```r
x[] <- user()
dim(x) <- user()
```

which will indicate that `x` is to be supplied by the user, and that
the _size_ of `x` will be inferred from the user-supplied values.  It
is untested, but this should work for 2d and 3d arrays too, but would
look like:

```r
x_2d[,] <- user()
x_3d[,,] <- user()
dim(x_2d) <- user()
dim(x_3d) <- user()
```

We don't have an assertion framework, but when that is in place it should
be possible to assert that the dimensions fall within a range.