Support for time dependent initial conditions
I don't know (and don't test yet) that this works sensibly for
array equations.  I think that most of the machinery is correct
but without tests it's a guess.

The other corner case to deal with is delay equations referenced
in the initial conditions.  These are flagged for now (and will
throw) but probably won't be terrible to implement.