Three argument delay
This allows for Berkeley Madonna-style specification of default
values for delays.  The third argument should be able to be
anything really, and it will be resolved so that the delay appears
after it in the graph.  This should allow for arrays, time dependent
variables etc, all to appear there which is cool.