Support multiple array variables
Requires some auxillary variables for "offsets".  The same approach
will be taken here for the auxillary dimensions which cause so much
trouble in the generation.  This is a bit of an annoyance with patching
the current internal structure on the way to creating the IR, but
doing this "properly" (i.e., in the parse) would definitely affect
existing odin generation