Start separating out array allocation and use
This turns out to be an unexpected bit of faff from earlier - I
think that there we generated code into two separate stages
previously (the user/creation stage and the dynamic stage), but
here we're going to continue expanding out the set of equations
to include an explicit allocation phase.

There is some work here to harmonise this with the interpolation
allocation, and there will be a similar allocation required for
the ring buffer(s) when getting delay models working.