Corner case for extent_min
This should be done in a more sensible way I think - we have different
schemas for the range and non-range versions of array indices, but
of course different range-nesses are possible for different indices.

So probably we should make this a array/array/map rather than array/map/array
as that will be easier to validate.