Include array dimension names in ir
This will take a while to filter through to the generation code
but it should completely remove to the need to use array_dim_name
there at all (and then lead to much simpler and more self contained
code generation).  There are lots of places where dimension names
are used in bulk, and this moves that responsibility here to be
run once then used mulitiple times.