Remove unused "stochastic" flag
It's not currently used in any generation and it would be
simple enough to add back in if needed.  I think that this is
something that might be important in the parse (working out that
something is implicitly time-dependent) but not important after
that sort of thing has been worked out).