All naming conventions now in ir
This removes the last usage of array_dim_name, which means that
the IR no longer depends on implicit naming conventions - it
is starting to become more self sufficient.  The closure approach
in the previous commit was needed to make this work nicely because
we're passing now the entire 'data' element into the rewriter.