Start generating IR for the delay case
This appears to be a minimal set of requirements.  Because different
DDE solvers will have different requirements for interrogating the
previous state of the system let's not be totally prescriptive about
how lookup is done there and delegate that to information already
held about the fundamental properties of the system.